refactor(painters/dom): extract paragraph fragment rendering into paragraph/ module
Moves paragraph fragment rendering, block versioning, and PM attribute
application out of renderer.ts into dedicated paragraph/ modules:

- paragraph/renderParagraphFragment.ts — fragment rendering entry point
- paragraph/block-version.ts — block version derivation + table-version hash
- paragraph/frame.ts — PM attribute application on the paragraph element

Trims renderer.ts by ~490 lines with no behavior change.
